With options to keep lubuntu desktop from getting get purged.

+++ script/minimize.sh
@@ -12,16 +12,23 @@ echo "==> Removing development packages"
 dpkg --list | awk '{ print $2 }' | grep -- '-dev$' | xargs apt-get -y purge
 echo "==> Removing documentation"
 dpkg --list | awk '{ print $2 }' | grep -- '-doc$' | xargs apt-get -y purge
-echo "==> Removing development tools"
-#dpkg --list | grep -i compiler | awk '{ print $2 }' | xargs apt-get -y purge
-#apt-get -y purge cpp gcc g++ 
-apt-get -y purge build-essential
+
+if [[ $PURGE_BUILD_ESSENTIAL  =~ true || $PURGE_BUILD_ESSENTIAL =~ 1 || $PURGE_BUILD_ESSENTIAL =~ yes ]]; then
+    echo "==> Removing development tools"
+    #dpkg --list | grep -i compiler | awk '{ print $2 }' | xargs apt-get -y purge
+    #apt-get -y purge cpp gcc g++ 
+    apt-get -y purge build-essential
+fi
 echo "==> Removing default system Ruby"
 apt-get -y purge ruby ri doc
 echo "==> Removing default system Python"
 apt-get -y purge python-dbus libnl1 python-smartpm python-twisted-core libiw30 python-twisted-bin libdbus-glib-1-2 python-pexpect python-pycurl python-serial python-gobject python-pam python-openssl libffi5
-echo "==> Removing X11 libraries"
-apt-get -y purge libx11-data xauth libxmuu1 libxcb1 libx11-6 libxext6
+
+if [[ $PURGE_X11  =~ true || $PURGE_X11 =~ 1 || $PURGE_X11 =~ yes ]]; then
+    echo "==> Removing X11 libraries"
+    apt-get -y purge libx11-data xauth libxmuu1 libxcb1 libx11-6 libxext6
+fi
+
 echo "==> Removing obsolete networking components"
 apt-get -y purge ppp pppconfig pppoeconf
 echo "==> Removing other oddities"
